requirements.txt
discord.py-self
requests
python-dotenv